refactor(node): separating internal chunk from register store implementation layer
