+++ sn_interface/src/messaging/data/register.rs
@@ -13,16 +13,12 @@ use crate::messaging::{data::OperationId, SectionAuth};
 use crate::types::register::Register;
 use crate::types::{
     register::{Entry, EntryHash, Policy, RegisterOp, User},
-    utils::serialise,
     RegisterAddress,
 };
 use tiny_keccak::{Hasher, Sha3};
 
 use serde::{Deserialize, Serialize};
 use xor_name::XorName;
-
-/// Deterministic Id for a register Cmd, takes into account the underlying cmd, and all sigs
-pub type RegisterCmdId = String;
 
 /// [`Register`] read operations.
 #[allow(clippy::large_enum_variant)]
@@ -244,20 +240,6 @@ impl RegisterCmd {
     /// This is not a unique identifier.
     pub fn name(&self) -> XorName {
         *self.dst_address().name()
-    }
-
-    /// Gets an operation id, deterministic for a RegisterCmd, it takes
-    /// the full Cmd and all signers into consideration
-    pub fn register_operation_id(&self) -> Result<RegisterCmdId> {
-        let mut hasher = Sha3::v256();
-
-        let bytes = serialise(&self).map_err(|_| Error::CouldNotSerialiseCmd)?;
-        let mut output = [0; 64];
-        hasher.update(&bytes);
-        hasher.finalize(&mut output);
-
-        let id = hex::encode(output);
-        Ok(id)
     }
 
     /// Returns the dst address of the register.
